refactor: update error tracking to monitoring
