+++ packages/analyzer/src/utils/analyze-feature.ts
@@ -9,7 +9,7 @@ import {
   stylingLibraries,
   stylingProcessor,
   transpiler,
-  errorTracking,
+  monitoring,
   uiLibrary,
   stateManagement,
   dates,
@@ -29,7 +29,7 @@ export const detectionMap = {
   router: router.detect,
   bundler: bundler.detect,
   transpiler: transpiler.detect,
-  errorTracking: errorTracking.detect,
+  monitoring: monitoring.detect,
   minifier: minifier.detect,
   stylingProcessor: stylingProcessor.detect,
   modules: modules.detect,
