+++ packages/features/src/categories/monitoring/meta/icons/newrelic.tsx
