+++ packages/features/src/categories/monitoring/meta/icons/raygun.tsx
