+++ packages/features/src/categories/monitoring/meta/icons/rollbar.tsx
