+++ packages/features/src/categories/monitoring/meta/icons/sentry.tsx
