+++ packages/features/src/categories/monitoring/patterns/index.ts
