+++ packages/features/src/categories/monitoring/patterns/newrelic.ts
