+++ packages/features/src/categories/monitoring/patterns/sentry.ts
