+++ packages/features/src/types/analysis.ts
@@ -25,7 +25,7 @@ export type AnalysisFeatures = {
   dates: AnalysisResult<typeof allFeatures.dates.detect>;
   stylingLibraries: AnalysisResult<typeof allFeatures.stylingLibraries.detect>;
   transpiler: AnalysisResult<typeof allFeatures.transpiler.detect>;
-  errorTracking: AnalysisResult<typeof allFeatures.errorTracking.detect>;
+  monitoring: AnalysisResult<typeof allFeatures.monitoring.detect>;
   platform: AnalysisResult<typeof allFeatures.platform.detect>;
   analytics: AnalysisResult<typeof allFeatures.analytics.detect>;
 };
